Enhancements and fixes for #27. BREAKING CHANGE: GetWindowLong and SetWindowLong now require WindowLongFlags value instead of a uint. There was no reason to leave open.

+++ WIndows.Forms/Extensions/ControlExtension.cs
@@ -1,10 +1,10 @@
 ﻿using System;
-using System.Runtime.InteropServices;
 using System.Windows.Forms;
 using static Vanara.PInvoke.User32_Gdi;
 
 namespace Vanara.Extensions
 {
+	/// <summary>Control extension methods.</summary>
 	public static partial class ControlExtension
 	{
 		/// <summary>
@@ -124,7 +124,7 @@ namespace Vanara.Extensions
 		/// <summary>Retrieves the window styles.</summary>
 		/// <param name="ctrl">The control.</param>
 		/// <returns>The window styles</returns>
-		public static int GetStyle(this Control ctrl) => GetWindowLongAuto(ctrl.Handle, (int)WindowLongFlags.GWL_STYLE).ToInt32();
+		public static int GetStyle(this Control ctrl) => GetWindowLongAuto(ctrl.Handle, WindowLongFlags.GWL_STYLE).ToInt32();
 
 		/// <summary>Removes the mnemonic, if one exists, from the string.</summary>
 		/// <param name="str">The string.</param>
@@ -175,11 +175,11 @@ namespace Vanara.Extensions
 		public static void SetStyle(this Control ctrl, int style, bool on = true)
 		{
 			var href = ctrl.Handle;
-			int oldstyle = GetWindowLongAuto(href, (int)WindowLongFlags.GWL_STYLE).ToInt32();
+			int oldstyle = GetWindowLongAuto(href, WindowLongFlags.GWL_STYLE).ToInt32();
 			if ((oldstyle & style) != style && on)
-				SetWindowLong(href, (int)WindowLongFlags.GWL_STYLE, new IntPtr(oldstyle | style));
+				SetWindowLong(href, WindowLongFlags.GWL_STYLE, new IntPtr(oldstyle | style));
 			else if ((oldstyle & style) == style && !on)
-				SetWindowLong(href, (int)WindowLongFlags.GWL_STYLE, new IntPtr(oldstyle & ~style));
+				SetWindowLong(href, WindowLongFlags.GWL_STYLE, new IntPtr(oldstyle & ~style));
 			ctrl.Refresh();
 		}
 	}
